creates navbar component
