--- frontend/app/dashboard/components/NavBar.tsx
@@ -1,16 +1,0 @@
-'use client'
-
-import Link from 'next/link'
-import { MENUS } from '../constants'
-
-export const NavBar: React.FC = () => {
-  return (
-    <ul>
-      {MENUS.map(({ label, path }) => (
-        <li key={label}>
-          <Link href={path}>{label}</Link>
-        </li>
-      ))}
-    </ul>
-  )
-}
