--- frontend/app/dashboard/constants/index.ts
@@ -1,5 +1,0 @@
-export const MENUS = [
-  { label: 'home', path: '*' },
-  { label: 'search', path: '*' },
-  { label: 'reading list', path: '*' }
-]
