+++ frontend/app/dashboard/layout.tsx
@@ -1,4 +1,6 @@
-import { NavBar } from './components/NavBar'
+'use client'
+
+import { NavBar } from 'components/navigations/NavBar'
 
 export default function DashboardLayout({ children }: { children: React.ReactNode }) {
   return (
