+++ frontend/app/layout.tsx
@@ -4,7 +4,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
   return (
     <html>
       <head />
-      <body>{children}</body>
+      <body className="bg-background">{children}</body>
     </html>
   )
 }
