+++ frontend/app/login/page.tsx
@@ -9,7 +9,7 @@ const queryClient = new QueryClient()
 export default function LoginPage() {
   return (
     <QueryClientProvider client={queryClient}>
-      <div className="bg-background flex h-screen w-full justify-center">
+      <div className="flex h-screen w-full justify-center">
         <div className="flex w-full max-w-sm flex-col items-center justify-center px-4">
           <div className="py-5">
             <Image width={200} height={200} alt="bookney-logo" src="/bookney-light.png" />
