+++ frontend/components/navigations/constants/index.ts
@@ -1,0 +1,5 @@
+export const MENUS = [
+  { label: 'home', path: '#' },
+  { label: 'search', path: '#' },
+  { label: 'reading list', path: '#' }
+]
